% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{acyl_cain_mass}
\alias{acyl_cain_mass}
\title{Acyl Chain Neutral Mass}
\usage{
acyl_cain_mass(carbon, db_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{db_bond}{Integer. Double bond count}
}
\value{
Numeric. Acyl chain neutral mass
}
\description{
Calculates neutral acyl chain mass (RCO- without OH).
C# equivalent: LipidMsmsCharacterizationUtility.acylCainMass
}
\keyword{internal}
