% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{acyl_chain_mass}
\alias{acyl_chain_mass}
\title{Calculate Acyl Chain Neutral Mass}
\usage{
acyl_chain_mass(carbon, double_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{double_bond}{Integer. Double bond count}
}
\value{
Numeric. Neutral mass of acyl chain (CnH(2n-1-2db)O)
}
\description{
Calculates neutral mass of acyl chain (R-CO- group) from chain composition.
Translation of C# acylCainMass().
}
\keyword{internal}
