% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{calcCiS}
\alias{calcCiS}
\title{Title}
\usage{
calcCiS(mSet, EIC = EIC, corval = 0.75, pval = 0.05, psg_list = NULL)
}
\arguments{
\item{mSet}{}

\item{EIC}{}

\item{corval}{}

\item{pval}{}

\item{psg_list}{}
}
\value{

}
\description{
Title
}
