% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_import.R
\name{CentroidMSData}
\alias{CentroidMSData}
\title{CentroidMSData}
\usage{
CentroidMSData(InFolder, OutFolder = tempdir(), ncore = 1)
}
\arguments{
\item{InFolder}{single file/folder name}

\item{OutFolder}{output folder name, if not exits, will create one}

\item{ncore}{the core number for parallel processing, default is 1}
}
\value{
will output a centroid mzML file into the input path
}
\description{
Convert the MS data as centroid
}
\examples{
InFolder <- system.file("mzData", package = "mtbls2")
# CentroidMSData(InFolder) #remove the # befroe your testing
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
