% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{controller.modifier}
\alias{controller.modifier}
\title{operators_1 : SetPeakParam ~ PerformParamsOptimization;
operators_2 : data_folder_trainning ~ PerformDataTrimming;
operators_3 : data_folder_sample ~ peak_profiling;
operators_4 : PerformPeakProfiling ~ PerformPeakAnnotation;
others_1 c1 : SetPeakParam;
others_1 c2 : ImportRawData - reading part;
others_1 c3 : ImportRawData - plotting part;
others_1 c4 : Start to do annotation.}
\usage{
controller.modifier(new_command, last_command, plan)
}
\description{
operators_1 : SetPeakParam ~ PerformParamsOptimization;
operators_2 : data_folder_trainning ~ PerformDataTrimming;
operators_3 : data_folder_sample ~ peak_profiling;
operators_4 : PerformPeakProfiling ~ PerformPeakAnnotation;
others_1 c1 : SetPeakParam;
others_1 c2 : ImportRawData - reading part;
others_1 c3 : ImportRawData - plotting part;
others_1 c4 : Start to do annotation.
}
