% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_params.R
\name{estimateSNThresh}
\alias{estimateSNThresh}
\title{Title}
\usage{
estimateSNThresh(no_match, sortedAllEIC, approvedPeaks)
}
\arguments{
\item{no_match}{}

\item{sortedAllEIC}{}

\item{approvedPeaks}{}
}
\value{

}
\description{
Title
}
