% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{Export.Annotation}
\alias{Export.Annotation}
\title{Export.Annotation}
\usage{
Export.Annotation(mSet = NULL, path = getwd())
}
\arguments{
\item{mSet}{mSet object, processed by FormatPeakList.}

\item{path}{character, used to specify the path for result rds and csv file. Default is the working directory.}
}
\description{
Export.Annotation is used to export the result of annotation
}
\examples{
data(mSet)
Export.Annotation(mSet, path = tempdir())
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jeff Xia \email{jeff.xia@mcgill.ca}
}
