% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{Export.PeakTable}
\alias{Export.PeakTable}
\title{Export.PeakTable}
\usage{
Export.PeakTable(mSet = NULL, path = getwd())
}
\arguments{
\item{mSet}{mSet object, processed by FormatPeakList.}

\item{path}{character, used to specify the path for result rds and csv file. Default is the working directory.#'}
}
\description{
Export.PeakTable is used to export the table of peak
}
\examples{
data(mSet)
Export.PeakTable(mSet, path = tempdir())
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jeff Xia \email{jeff.xia@mcgill.ca}
}
