% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{fatty_acid_product_ion}
\alias{fatty_acid_product_ion}
\title{Fatty Acid Product Ion m/z (RCOO-)}
\usage{
fatty_acid_product_ion(carbon, db_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{db_bond}{Integer. Double bond count}
}
\value{
Numeric. m/z of RCOO- ion
}
\description{
Calculates the m/z of a fatty acid carboxylate anion (RCOO-).
C# equivalent: LipidMsmsCharacterizationUtility.fattyacidProductIon
}
\keyword{internal}
