% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize_params.R
\name{filterPpmError}
\alias{filterPpmError}
\title{Title}
\usage{
filterPpmError(
  approvedPeaks,
  useGap,
  varExpThresh,
  returnPpmPlots,
  plotDir,
  observedPeak,
  filename
)
}
\arguments{
\item{approvedPeaks}{}

\item{useGap}{}

\item{varExpThresh}{}

\item{returnPpmPlots}{}

\item{plotDir}{}

\item{observedPeak}{}

\item{filename}{}
}
\value{

}
\description{
Title
}
