% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Infusion_function.R
\name{FormatInfusionFeatureTable}
\alias{FormatInfusionFeatureTable}
\title{FormatInfusionFeatureTable}
\usage{
FormatInfusionFeatureTable(mSet, minFraction = 0.1, minSamples = 0)
}
\arguments{
\item{mSet}{mSet Object containing detected infusion features.}

\item{minFraction}{Numeric in between 0 and 1, minimum fraction for feature grouping.}

\item{minSamples}{Integer, minimum number of samples required.}
}
\value{
mSet Object
}
\description{
FormatInfusionFeatureTable
}
\examples{
todo
}
