% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{FormatMSnAnnotation}
\alias{FormatMSnAnnotation}
\title{FormatMSnAnnotation}
\usage{
FormatMSnAnnotation(mSet = NULL, topN = 5L, isLipidomics = FALSE)
}
\arguments{
\item{mSet}{mSet Object containing exported annotation results.}

\item{topN}{Integer, number of top-ranked annotations to keep.}

\item{isLipidomics}{Logical, whether to format output for lipidomics workflows.}
}
\value{
mSet Object
}
\description{
FormatMSnAnnotation
}
