% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{FormatPeakList}
\alias{FormatPeakList}
\title{Format Peak List}
\usage{
FormatPeakList(
  mSet,
  annParams,
  filtIso = TRUE,
  filtAdducts = FALSE,
  missPercent = 0.75
)
}
\arguments{
\item{mSet}{The mSet object generated by the PerformPeakAnnotation function.}

\item{annParams}{The object created using the SetAnnotationParam function,
containing user's specified or default parameters for downstream
raw MS data pre-processing.}

\item{filtIso}{Logical, filter out all isotopes except for \verb{[M]}+ for
positive ion mode and \verb{[M]}- for negative ion mode. By default it is
set to true.}

\item{filtAdducts}{Logical, filter out all adducts except \verb{[M+H]}+ for
positive ion more and \verb{[M-H]}- for negative ion mode. By default it is set to false.}

\item{missPercent}{Numeric, specify the threshold to remove features
missing in X\\% of samples. For instance, 0.5 specifies to remove features
that are missing from 50\\% of all samples per group. Method is only valid
when there are two groups.}
}
\description{
This function formats the CAMERA output to a usable format for OptiLCMS.
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
annParams <- SetAnnotationParam(polarity = 'positive',
                                mz_abs_add = 0.035);

## Perform peak annotation with newly deinfed annParams
# mSet <- PerformPeakAnnotation(mSet = mSet,
#                               annotaParam = annParams,
#                               ncore =1)
## Format the PeakList
mSet <- FormatPeakList(mSet = mSet,
                       annParams,
                       filtIso =FALSE,
                       filtAdducts = FALSE,
                       missPercent = 1)
}
\seealso{
\code{\link{ExecutePlan}} and \code{\link{PerformPeakProfiling}} for the whole pipeline.
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
