% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{GaussModel}
\alias{GaussModel}
\title{GaussModel}
\usage{
GaussModel(x, mu, sigma, h)
}
\arguments{
\item{x}{a numeric vector of values at which to evaluate the model}

\item{mu}{mean  of the distribution function}

\item{sigma}{standard deviation of the distribution fuction}

\item{h}{height of the distribution function}
}
\description{
GaussModel
}
