% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_acylhexceramide_molecule_obj_as_level2}
\alias{get_acylhexceramide_molecule_obj_as_level2}
\title{Create AcylHexCeramide (3-chain) Level 2 Molecule Object}
\usage{
get_acylhexceramide_molecule_obj_as_level2(
  lipid_class,
  lbm_class,
  hydroxy_string,
  sph_carbon,
  sph_double,
  acyl_carbon,
  acyl_double,
  ester_carbon,
  ester_double,
  score = 0,
  acyl_oxidized = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{sph_carbon, sph_double}{Integer. Sphingoid base}

\item{acyl_carbon, acyl_double}{Integer. Amide-linked acyl chain}

\item{ester_carbon, ester_double}{Integer. O-linked ether acyl chain}

\item{score}{Numeric.}

\item{acyl_oxidized}{Integer. Oxidation count on acyl chain}
}
\value{
LipidMolecule list
}
\description{
Translation of C# getAcylhexceramideMoleculeObjAsLevel2().
}
\keyword{internal}
