% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_acylhexceramide_molecule_obj_as_level2_0}
\alias{get_acylhexceramide_molecule_obj_as_level2_0}
\title{Create AcylHexCeramide (2-chain, ceramide bulk + ether) Level 2 Molecule Object}
\usage{
get_acylhexceramide_molecule_obj_as_level2_0(
  lipid_class,
  lbm_class,
  hydroxy_string,
  ceramide_carbon,
  ceramide_double,
  ester_carbon,
  ester_double,
  score = 0,
  acyl_oxidized = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{ceramide_carbon, ceramide_double}{Integer. Combined ceramide carbons/db}

\item{ester_carbon, ester_double}{Integer. O-linked ether acyl chain}

\item{score}{Numeric.}

\item{acyl_oxidized}{Integer. Oxidation count}
}
\value{
LipidMolecule list
}
\description{
Translation of C# getAcylhexceramideMoleculeObjAsLevel2_0().
}
\keyword{internal}
