% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_asm_molecule_obj_as_level2}
\alias{get_asm_molecule_obj_as_level2}
\title{Create AcylSphingomyelin (ASM, 3-chain) Level 2 Molecule Object}
\usage{
get_asm_molecule_obj_as_level2(
  lipid_class,
  lbm_class,
  hydroxy_string,
  sph_carbon,
  sph_double,
  acyl_carbon,
  acyl_double,
  ester_carbon,
  ester_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{sph_carbon, sph_double}{Integer. Sphingoid base}

\item{acyl_carbon, acyl_double}{Integer. Amide-linked acyl chain}

\item{ester_carbon, ester_double}{Integer. FA ester chain}

\item{score}{Numeric.}
}
\value{
LipidMolecule list
}
\description{
Translation of C# getAsmMoleculeObjAsLevel2().
}
\keyword{internal}
