% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_asm_molecule_obj_as_level2_0}
\alias{get_asm_molecule_obj_as_level2_0}
\title{Create AcylSphingomyelin (ASM, 2-chain bulk) Level 2 Molecule Object}
\usage{
get_asm_molecule_obj_as_level2_0(
  lipid_class,
  lbm_class,
  hydroxy_string,
  cer_carbon,
  cer_double,
  ester_carbon,
  ester_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{cer_carbon, cer_double}{Integer. Combined ceramide carbons/db}

\item{ester_carbon, ester_double}{Integer. FA ester chain}

\item{score}{Numeric.}
}
\value{
LipidMolecule list
}
\description{
Translation of C# getAsmMoleculeObjAsLevel2_0().
}
\keyword{internal}
