% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_esterceramide_molecule_obj_as_level2}
\alias{get_esterceramide_molecule_obj_as_level2}
\title{Create Ester Ceramide (3-chain, fully resolved) Level 2 Molecule Object}
\usage{
get_esterceramide_molecule_obj_as_level2(
  lipid_class,
  lbm_class,
  hydroxy_string,
  sph_carbon,
  sph_double,
  acyl_carbon,
  acyl_double,
  ester_carbon,
  ester_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{sph_carbon, sph_double}{Integer. Sphingoid base chain}

\item{acyl_carbon, acyl_double}{Integer. Amide-linked acyl chain}

\item{ester_carbon, ester_double}{Integer. Ester-linked acyl chain}

\item{score}{Numeric.}
}
\value{
LipidMolecule list
}
\description{
Handles Cer_EOS, Cer_EODS, HexCer_EOS, Cer_EBDS etc.
Translation of C# getEsterceramideMoleculeObjAsLevel2().
}
\keyword{internal}
