% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{get_esterceramide_molecule_obj_as_level2_1}
\alias{get_esterceramide_molecule_obj_as_level2_1}
\title{Create Ester Ceramide (2-chain, sph + hydroxy-acyl) Level 2 Molecule Object}
\usage{
get_esterceramide_molecule_obj_as_level2_1(
  lipid_class,
  lbm_class,
  hydroxy_string,
  sph_carbon,
  sph_double,
  acyl_carbon,
  acyl_double,
  score = 0
)
}
\arguments{
\item{lipid_class}{Character.}

\item{lbm_class}{Character.}

\item{hydroxy_string}{Character. "m","d","t"}

\item{sph_carbon, sph_double}{Integer. Sphingoid base}

\item{acyl_carbon, acyl_double}{Integer. Hydroxy-acyl chain}

\item{score}{Numeric.}
}
\value{
LipidMolecule list
}
\description{
Translation of C# getEsterceramideMoleculeObjAsLevel2_1().
}
\keyword{internal}
