% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_import.R
\name{ImportRawMSData}
\alias{ImportRawMSData}
\title{Import raw MS data}
\usage{
ImportRawMSData(
  mSet = NULL,
  path = getwd(),
  metadata = NULL,
  mode = "onDisk",
  plotSettings = SetPlotParam(),
  running.controller = NULL
)
}
\arguments{
\item{mSet}{mSet Object, can be optional. Usually generated by InitDataObjects("spec", "raw", FALSE) before the data import.}

\item{path}{Character, input the path to the folder containing
the raw MS spectra to be processed. Or a character vector containing all raw files absolute paths.}

\item{metadata}{Data.frame or character. A phenotype data frame or a absolute path of the metadata file (.txt) for all samples, optional.
In the option, first column should be the sample name, while second column is the corresponding group name. If ommited, all samples in the same sub-folder will be
considered as one group.}

\item{mode}{Character, the data input mode. Default is "onDisk" to avoid memory crash. "inMemory" will
read data into memory.}

\item{plotSettings}{List, plotting parameters produced by SetPlotParam Function. "plot.opts" can be added through this
function for samples numbers for plotting. Defalut is "default", "all" will apply all samples for plotting and may cause
memory crash, especially for large sample dataset.}

\item{running.controller}{The resuming pipeline running controller. Optional. Don't need to define by hand.}
}
\description{
This function handles the reading in of
raw MS data (.mzML, .CDF and .mzXML). Users must set
their working directory to the folder containing their raw
data, divided into two subfolders named their desired group labels. The
function will output two chromatograms into the user's working directory, a
base peak intensity chromatogram (BPIC) and a total ion
chromatogram (TIC). Further, this function sets the number of cores
to be used for parallel processing. It first determines the number of cores
within a user's computer and then sets it that number/2.
}
\examples{
##' Load OptiLCMS package
library(OptiLCMS)
##' Get raw spectra files
DataFiles <- dir(system.file("mzData", package = "mtbls2"), full.names = TRUE,
                 recursive = TRUE)[c(10:12, 14:16)]
##' Create a phenodata data.frame
pd <- data.frame(sample_name = sub(basename(DataFiles), pattern = ".mzData",
                                   replacement = "", fixed = TRUE),
                 sample_group = c(rep("col0", 3), rep("cyp79", 3)),
                 stringsAsFactors = FALSE)
##' Import raw spectra
mSet <- ImportRawMSData(path = DataFiles, metadata = pd);
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
