% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{InitializaPlan}
\alias{InitializaPlan}
\title{Initializing running plan}
\usage{
InitializaPlan(type = "spec", path)
}
\arguments{
\item{type}{}

\item{path}{}
}
\description{
Initializing running plan
}
