% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{is_diagnostic_fragment_exist}
\alias{is_diagnostic_fragment_exist}
\title{Check for Diagnostic Fragment in Spectrum}
\usage{
is_diagnostic_fragment_exist(spectrum, ms2_tolerance, diagnostic_mz, threshold)
}
\arguments{
\item{spectrum}{List-of-lists or data frame with Mass and Intensity columns}

\item{ms2_tolerance}{Numeric. MS2 mass tolerance in Da}

\item{diagnostic_mz}{Numeric. Target m/z for diagnostic fragment}

\item{threshold}{Numeric. Minimum intensity threshold}
}
\value{
Logical. TRUE if diagnostic fragment found
}
\description{
Check for Diagnostic Fragment in Spectrum
}
\keyword{internal}
