% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{is_fragment1_greater_than_fragment2}
\alias{is_fragment1_greater_than_fragment2}
\title{Compare Intensity of Two Fragments}
\usage{
is_fragment1_greater_than_fragment2(
  spectrum,
  ms2_tolerance,
  fragment1,
  fragment2
)
}
\arguments{
\item{spectrum}{List-of-lists or data frame with Mass and Intensity columns}

\item{ms2_tolerance}{Numeric. Tolerance for m/z matching}

\item{fragment1}{Numeric. m/z of first fragment}

\item{fragment2}{Numeric. m/z of second fragment}
}
\value{
Logical. TRUE if fragment1 intensity > fragment2 intensity
}
\description{
Compare Intensity of Two Fragments
}
\keyword{internal}
