% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_case_species}
\alias{judge_if_case_species}
\title{Judge If CASE Species}
\usage{
judge_if_case_species(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  adduct
)
}
\description{
Judge If CASE Species
}
\keyword{internal}
