% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_ceramideahd9}
\alias{judge_if_ceramideahd9}
\title{Judge Cer-AH-d9 (CeramideAhD9)}
\usage{
judge_if_ceramideahd9(
  spectrum,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  min_sph_carbon,
  max_sph_carbon,
  min_sph_double_bond,
  max_sph_double_bond,
  adduct
)
}
\arguments{
\item{spectrum}{MS/MS spectrum data}

\item{ms2_tolerance}{MS/MS mass tolerance}

\item{theoretical_mz}{Theoretical m/z}

\item{total_carbon}{Total carbon atoms}

\item{total_double_bond}{Total double bonds}

\item{min_sph_carbon}{Min sphingoid carbon}

\item{max_sph_carbon}{Max sphingoid carbon}

\item{min_sph_double_bond}{Min sphingoid double bonds}

\item{max_sph_double_bond}{Max sphingoid double bonds}

\item{adduct}{Adduct details (IonMode, AdductIonName)}
}
\value{
LipidMolecule object or NULL
}
\description{
Identifies deuterated ceramide with alpha-hydroxylation (Cer-AH-d9).
Same logic as Cer-AH but with d9 (deuterium) labeling and adjusted mass calculations.
Includes D2 mass substitutions in acyl fragments.
}
