% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_etherpi}
\alias{judge_if_etherpi}
\title{Judge If EtherPI}
\usage{
judge_if_etherpi(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  min_sn_carbon,
  max_sn_carbon,
  min_sn_double_bond,
  max_sn_double_bond,
  adduct
)
}
\description{
Judge If EtherPI
}
\keyword{internal}
