% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_gd1b}
\alias{judge_if_gd1b}
\title{Judge If GD1b}
\usage{
judge_if_gd1b(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  min_sph_carbon,
  max_sph_carbon,
  min_sph_double_bond,
  max_sph_double_bond,
  adduct
)
}
\description{
Judge If GD1b
}
\keyword{internal}
