% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_gm3}
\alias{judge_if_gm3}
\title{Judge If GM3 (Ganglioside GM3)}
\usage{
judge_if_gm3(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  min_sph_carbon,
  max_sph_carbon,
  min_sph_double_bond,
  max_sph_double_bond,
  adduct
)
}
\arguments{
\item{ms_scan_prop}{List containing 'Spectrum'}

\item{ms2_tolerance}{Numeric. MS2 mass tolerance}

\item{theoretical_mz}{Numeric. Theoretical m/z}

\item{total_carbon}{Integer. Total carbons}

\item{total_double_bond}{Integer. Total double bonds}

\item{min_sph_carbon}{Integer. Minimum sphingoid carbons}

\item{max_sph_carbon}{Integer. Maximum sphingoid carbons}

\item{min_sph_double_bond}{Integer. Minimum sphingoid double bonds}

\item{max_sph_double_bond}{Integer. Maximum sphingoid double bonds}

\item{adduct}{List. Adduct information}
}
\value{
List representing LipidMolecule or NULL
}
\description{
Identifies GM3 gangliosides based on MS/MS fragmentation patterns.
Supports both negative and positive ion modes.
}
\keyword{internal}
