% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_phytosphingosine}
\alias{judge_if_phytosphingosine}
\title{Judge If Phytosphingosine}
\usage{
judge_if_phytosphingosine(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  adduct
)
}
\description{
Judge If Phytosphingosine
}
\keyword{internal}
