% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_sphinganine}
\alias{judge_if_sphinganine}
\title{Judge If Sphinganine}
\usage{
judge_if_sphinganine(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  adduct
)
}
\description{
Judge If Sphinganine
}
\keyword{internal}
