% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_sphingosine}
\alias{judge_if_sphingosine}
\title{Judge If Sphingosine}
\usage{
judge_if_sphingosine(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  adduct
)
}
\description{
Judge If Sphingosine
}
\keyword{internal}
