% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{lipidmsmscharacterization_stub_fatty_acids_others}
\alias{lipidmsmscharacterization_stub_fatty_acids_others}
\title{Judgment Functions Stub - Fatty Acids & Others}
\description{
\itemize{
\item judge_if_fatty_acid (FA)
\item judge_if_acylcarnitine (CAR)
\item judge_if_cholesteryl_ester (CE)
\item judge_if_wax_ester (WE)
\item judge_if_coenzyme_q, judge_if_vitamin_e, etc.
}
}
\keyword{internal}
