% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{lipidmsmscharacterization_stub_glycerolipids}
\alias{lipidmsmscharacterization_stub_glycerolipids}
\title{Judgment Functions Stub - Glycerolipids}
\description{
The following functions are skeletal implementations pending full translation:
\itemize{
\item judge_if_triacylglycerol (TAG/TG)
\item judge_if_dag (Diacylglycerol)
\item judge_if_mag (Monoacylglycerol)
\item judge_if_cardiolipin (CL, MLCL, DLCL)
}
}
\keyword{internal}
