% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{lipidmsmscharacterization_stub_phospholipids}
\alias{lipidmsmscharacterization_stub_phospholipids}
\title{Judgment Functions Stub - Phospholipids}
\description{
The following functions are skeletal implementations pending full translation:
\itemize{
\item judge_if_phosphatidylethanolamine (PE)
\item judge_if_phosphatidylserine (PS)
\item judge_if_phosphatidylglycerol (PG)
\item judge_if_phosphatidylinositol (PI)
\item judge_if_lysopc / judge_if_lysope (Lysophospholipids)
\item judge_if_ether_pc / judge_if_ether_pe (Ether lipids)
\item judge_if_oxpc / judge_if_oxpe / judge_if_oxpi / judge_if_oxps (Oxidized lipids)
}
}
\keyword{internal}
