% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{lipidmsmscharacterization_stub_sphingolipids}
\alias{lipidmsmscharacterization_stub_sphingolipids}
\title{Judgment Functions Stub - Sphingolipids}
\description{
The following ceramide variants and sphingomyelin functions pending:
\itemize{
\item judge_if_sphingomyelin (SM)
\item judge_if_ceramide_ns / judge_if_ceramide_nds (Ceramides)
\item judge_if_hex_ceramide / judge_if_hex2_ceramide
\item judge_if_acyl_sm (Acyl-sphingomyelin)
\item judge_if_ganglioside_gm1 / judge_if_ganglioside_gm3, etc.
}
}
\keyword{internal}
