% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_setting.R, R/Perform_functions.R
\name{MessageOutput}
\alias{MessageOutput}
\title{Title}
\usage{
MessageOutput(mes, ecol, progress)

MessageOutput(mes, ecol, progress)
}
\arguments{
\item{mes}{}

\item{ecol}{}

\item{progress}{}
}
\value{



}
\description{
Title

Title
}
