% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{mSet2xcmsSet}
\alias{mSet2xcmsSet}
\title{mSet2xcmsSet}
\usage{
mSet2xcmsSet(mSet)
}
\description{
mSet2xcmsSet
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
