% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiLCMS_Class.R
\docType{class}
\name{mSetRule-class}
\alias{mSetRule-class}
\title{This class is highly referenced from CAMERA but will improve later}
\description{
This class is highly referenced from CAMERA but will improve later
}
\references{
Kuhl C, Tautenhahn R, Boettcher C, Larson TR, Neumann S (2012).
"CAMERA: an integrated strategy for compound spectra extraction and annotation of
liquid chromatography/mass spectrometry data sets." Analytical Chemistry, 84, 283-289.
http://pubs.acs.org/doi/abs/10.1021/ac202450g.
}
