% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_function.R
\docType{package}
\name{OptiLCMS}
\alias{OptiLCMS}
\title{OptiLCMS: A package for computating the notorious bar statistic.}
\description{
The OptiLCMS package provides a pipeline for metabolomics processing.
}
\section{OptiLCMS functions}{

The OptiLCMS functions ...
}

