% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{ov_qs_exists}
\alias{ov_qs_exists}
\title{Check whether serialized data exists in \code{.qs2} or \code{.qs} form}
\usage{
ov_qs_exists(file)
}
\arguments{
\item{file}{A file path ending with \code{.qs2} or \code{.qs}.}
}
\value{
Logical scalar indicating whether either file exists.
}
\description{
Check whether serialized data exists in \code{.qs2} or \code{.qs} form
}
