% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{ov_qs_save}
\alias{ov_qs_save}
\title{Save serialized data via qs2}
\usage{
ov_qs_save(obj, file, ...)
}
\arguments{
\item{obj}{An R object to serialize.}

\item{file}{Output file path.}

\item{...}{Additional arguments passed to \code{qs2::qs_save}.}
}
\description{
Save serialized data via qs2
}
