% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PeakPicking_Massifquant_slave}
\alias{PeakPicking_Massifquant_slave}
\title{PeakPicking_Massifquant_slave}
\usage{
PeakPicking_Massifquant_slave(x, param)
}
\arguments{
\item{x}{ms objects}

\item{param}{parameters set for processing}
}
\description{
PeakPicking_Massifquant_slave
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
