% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PerformDataInspect}
\alias{PerformDataInspect}
\title{PerformDataInspect}
\usage{
PerformDataInspect(
  datapath = NULL,
  rt.range = c(0, 0),
  mz.range = c(0, 0),
  dimension = "3D",
  res = 100
)
}
\arguments{
\item{datapath}{Character, the path of the raw MS data files (.mzXML, .CDF and .mzML)
for the visual and intuitive data inspectation or the file folder (if only a folder path provided, the first file will
be inspected).}

\item{rt.range}{Numerics, a congregation of two values to define the lower and upper RT range (seconds) for
users to inspect. This is an optional parameter, if absent, will display the MS of the whole RT range.}

\item{mz.range}{Numerics, a congregation of two values to define the lower and upper mz range for
users to inspect. This is an optional parameter, if absent, will display the MS of the whole mz range.}

\item{dimension}{Character, the dimension for sample to display, including '2D' or '3D'. The default is '3D'.}

\item{res}{Numeric, the resolution for data inspectation. The larger the value, the higher the resolution.
The default value is 100. This value is usually clearly enough and also give consideration to the speed.}
}
\value{
will output a figure for viewing the data structure
}
\description{
This functions provide a path for users to visually inspect their raw data before the data
trimming so as to remove the dirty or significantly uneluted peaks.
}
\examples{
## Get raw spectra files
DataFiles <- dir(system.file("mzData", package = "mtbls2"), full.names = TRUE,
                 recursive = TRUE)[c(10:12, 14:16)]
PerformDataInspect(DataFiles[1])
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
