% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformDDADeconvolution}
\alias{PerformDDADeconvolution}
\title{PerformDDADeconvolution}
\usage{
PerformDDADeconvolution(
  mSet = NULL,
  ppm1 = 5,
  ppm2 = 15,
  sn = 12,
  filtering = 2000,
  window_size = 1,
  intensity_thresh = 1000,
  database_path = "",
  ncores = 1L,
  decoOn = TRUE,
  useEntropy = FALSE
)
}
\arguments{
\item{mSet}{mSet}

\item{ppm1}{ppm1}

\item{ppm2}{ppm2}

\item{sn}{sn}

\item{filtering}{filtering}

\item{window_size}{window_size}

\item{intensity_thresh}{intensity_thresh}

\item{database_path}{database_path}

\item{ncores}{ncores}

\item{decoOn}{decoOn}

\item{useEntropy}{useEntropy}
}
\value{
mSet Object
}
\description{
PerformDDADeconvolution
}
\examples{
to add
}
