% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformDIADeconvolution}
\alias{PerformDIADeconvolution}
\title{PerformDIADeconvolution}
\usage{
PerformDIADeconvolution(
  mSet = NULL,
  min_width = 5,
  ppm2,
  sn = 12,
  span = 0.3,
  filtering = 2000,
  ncores = 1L
)
}
\arguments{
\item{mSet}{mSet Object contains raw spectral data from \emph{PerformMSnImport}}

\item{min_width}{minimum peak width value, in seconds}

\item{ppm2}{Numeric, MS2 m/z tolerance in ppm.}

\item{sn}{Numeric, signal-to-noise threshold for MS2 deconvolution.}

\item{span}{Numeric, smoothing span used during chromatographic processing.}

\item{filtering}{Numeric, minimum intensity filter threshold.}

\item{ncores}{Integer, number of CPU cores to use.}
}
\value{
mSet Object
}
\description{
PerformDIADeconvolution
}
\examples{
to add
}
