% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Infusion_function.R
\name{PerformInusionFeatureDection}
\alias{PerformInusionFeatureDection}
\title{PerformInusionFeatureDection}
\usage{
PerformInusionFeatureDection(
  mSet,
  method = "cwt",
  noise = NA,
  snthresh = 3,
  ppm = 10,
  scales = c(1, 4, 9),
  groups = NULL,
  ncores = 1L
)
}
\arguments{
\item{mSet}{mSet Object containing raw infusion MS data.}

\item{method}{Character, peak detection method ("cwt" or "KDE").}

\item{noise}{Numeric, noise threshold for peak detection.}

\item{snthresh}{Numeric, signal-to-noise threshold.}

\item{ppm}{Numeric, mass accuracy in ppm.}

\item{scales}{Numeric vector, wavelet scales for CWT method.}

\item{ncores}{Integer, number of CPU cores to use.}
}
\value{
mSet Object
}
\description{
PerformInusionFeatureDection
}
\examples{
todo
}
