% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lipidomics_processing.R
\name{PerformLipidsDBSearchingBatch}
\alias{PerformLipidsDBSearchingBatch}
\title{PerformLipidsDBSearchingBatch}
\usage{
PerformLipidsDBSearchingBatch(
  mSet = NULL,
  ppm1 = 5,
  ppm2 = 15,
  rt_tol = 0,
  database_path =
    "/Volumes/ExtremeSSD/Lipidomics_projects/database_curation/lipidomics_complete.sqlite",
  use_rt = FALSE,
  ncores = 1L,
  useEntropy = FALSE
)
}
\description{
Batch database searching for lipidomics DB (sqlite). Stores results in mSet@MSnResults$LipidDBMatchRes
}
