% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformMirrorPlotting}
\alias{PerformMirrorPlotting}
\title{PerformMirrorPlotting}
\usage{
PerformMirrorPlotting(
  mSet = NULL,
  cutoff_relative = 5,
  ppm = 25,
  display_plot = FALSE,
  width = 8,
  height = 6,
  dpi = 300,
  format = "png",
  interactive = FALSE
)
}
\arguments{
\item{mSet}{mSet}

\item{cutoff_relative}{cutoff value for relative intensity}

\item{ppm}{ppm of mz of msms fragment}

\item{display_plot}{display mirror plot or not, TRUE or FALSE, default is FALSE}

\item{width}{width of the image, default is 8}

\item{height}{height of the image, default is 6}

\item{dpi}{dpi value, default is 300}

\item{interactive}{to make figure interactive, default is FALSE. TRUE or FALSE.}
}
\description{
PerformMirrorPlotting
}
