% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformMSnImport}
\alias{PerformMSnImport}
\title{PerformMSnImport}
\usage{
PerformMSnImport(
  mSet = NULL,
  filesPath = NULL,
  targetFeatures = NULL,
  acquisitionMode = "DDA",
  SWATH_file = NULL
)
}
\arguments{
\item{mSet}{mSet}

\item{filesPath}{filesPath}

\item{targetFeatures}{targetFeatures}

\item{acquisitionMode}{acquisitionMode}

\item{SWATH_file}{SWATH_file}
}
\value{
mSet Object
}
\description{
PerformMSnImport
}
\examples{
to add
}
