% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakAlignment}
\alias{PerformPeakAlignment}
\title{PerformPeakAlignment}
\usage{
PerformPeakAlignment(mSet)
}
\arguments{
\item{mSet}{the mSet object generated by PerformPeakPicking function.}
}
\description{
PerformPeakAlignment
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
