% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{PerformPeakAnnotation}
\alias{PerformPeakAnnotation}
\title{Perform peak annotation}
\usage{
PerformPeakAnnotation(mSet, annotaParam, ncore = 1, running.controller = NULL)
}
\arguments{
\item{mSet}{mSet object, usually generated by 'PerformPeakProfiling' here.}

\item{annotaParam}{The object created using the SetAnnotationParam function,
containing user's specified or default parameters for downstream
raw MS data pre-processing.}

\item{ncore}{annotation running core. Default is 1. Parallel running will be supported soon.}

\item{running.controller}{The resuming pipeline running controller. Optional. Don't need to define by hand.}
}
\value{
will return an mSet object wirh annotation finished
}
\description{
This function performs peak annotation on
the xset object created using the PerformPeakPicking function.
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
annParams <- SetAnnotationParam(polarity = 'positive',
                                mz_abs_add = 0.035);

## Perform peak annotation with newly deinfed annParams
# mSet <- PerformPeakAnnotation(mSet = mSet,
#                               annotaParam = annParams,
#                               ncore =1)
}
\references{
Kuhl C, Tautenhahn R, Boettcher C, Larson TR, Neumann S (2012).
"CAMERA: an integrated strategy for compound spectra extraction and annotation of
liquid chromatography/mass spectrometry data sets." Analytical Chemistry, 84, 283-289.
http://pubs.acs.org/doi/abs/10.1021/ac202450g.
}
\seealso{
\code{\link{ExecutePlan}} and \code{\link{PerformPeakProfiling}} for the whole pipeline.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
