% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakFiling}
\alias{PerformPeakFiling}
\title{PerformPeakFiling}
\usage{
PerformPeakFiling(mSet, BPPARAM = bpparam())
}
\arguments{
\item{mSet}{the mSet object generated by PerformPeakPicking function.}

\item{BPPARAM}{parallel method used for data processing. Default is bpparam().}

\item{param}{param list generated by updateRawSpectraParam function.}
}
\description{
PerformPeakFiling
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
