% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakGrouping}
\alias{PerformPeakGrouping}
\title{PerformPeakGrouping}
\usage{
PerformPeakGrouping(mSet)
}
\arguments{
\item{mSet}{mSet object.}
}
\description{
PerformPeakGrouping
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
