% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakPicking}
\alias{PerformPeakPicking}
\title{PerformPeakPicking}
\usage{
PerformPeakPicking(mSet, BPPARAM = bpparam())
}
\arguments{
\item{mSet}{the raw data object read by ImportRawMSData function.}

\item{BPPARAM}{parallel method used for data processing. Default is bpparam(). Optional.}
}
\value{
will return an mSet object with peaks picked
}
\description{
This funciton is used to Perform Peak Picking on an object generated by ImportRawMSdata
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
# mSet <- PerformPeakPicking(mSet);
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
