% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformResultsExport}
\alias{PerformResultsExport}
\title{PerformResultsExport}
\usage{
PerformResultsExport(
  mSet = NULL,
  type = 0L,
  topN = 10L,
  ncores = 1L,
  lipids = F
)
}
\arguments{
\item{mSet}{mSet Object containing DB search results.}

\item{type}{Integer export mode (0L, 1L, 2L, or 3L).}

\item{topN}{Integer, maximum number of annotations retained per feature.}

\item{ncores}{Integer, number of CPU cores to use.}

\item{lipids}{Logical, whether to apply lipid-class export formatting.}
}
\value{
mSet Object
}
\description{
PerformResultsExport
}
