% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformSpectrumConsenus}
\alias{PerformSpectrumConsenus}
\title{PerformSpectrumConsenus}
\usage{
PerformSpectrumConsenus(
  mSet = NULL,
  ppm2,
  concensus_fraction = 0.5,
  database_path = "",
  use_rt = FALSE,
  user_dbCorrection = TRUE,
  useEntropy = FALSE
)
}
\arguments{
\item{mSet}{mSet Object after MS2 deconvolution.}

\item{ppm2}{Numeric, MS2 m/z tolerance in ppm.}

\item{concensus_fraction}{Numeric in \link{0,1}, minimum fraction for consensus peaks.}

\item{database_path}{Character, path to the SQLite database used for correction.}

\item{use_rt}{Logical, whether to use retention time in consensus processing.}

\item{user_dbCorrection}{Logical, whether to apply database-based correction.}
}
\value{
mSet Object
}
\description{
PerformSpectrumConsenus
}
\examples{
to add
}
