% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROI_Extraction.R
\name{plot.MS_3D}
\alias{plot.MS_3D}
\title{Function for 3D ms plotting}
\usage{
\method{plot}{MS_3D}(object)
}
\description{
Function for 3D ms plotting (internal use only)
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}
}
