% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotBPIs}
\alias{plotBPIs}
\title{plotBPIs}
\usage{
plotBPIs(mSet = NULL, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mSet}{mSet Object, should be processed by ImportMSData.}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png". (only works for web version)}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72. (only works for web version)}

\item{filename}{Character, to give the filename for the TIC plotting.}
}
\description{
plotBPIs is used to plot the BPI of all files
}
