% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotSingleTIC}
\alias{plotSingleTIC}
\title{plotSingleTIC}
\usage{
plotSingleTIC(mSet = NULL, filename, imagename)
}
\arguments{
\item{mSet}{mSet Object, should be processed by ImportMSData.}

\item{filename}{Character, to give the filename for the TIC plotting.}

\item{imagename}{Character, to give the filename of the TIC plotted.}
}
\description{
plotSingleTIC
}
