% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PlotSpectraBPIadj}
\alias{PlotSpectraBPIadj}
\title{PlotSpectraBPIadj}
\usage{
PlotSpectraBPIadj(mSet, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mSet}{mSet object, usually generated after the peakannotaion finished here.}

\item{imgName}{Character, to give the name of BPI figures ploted.}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png".}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72.}

\item{width}{Numeric, to define the width of the figure. Height = width * 0.618.}
}
\description{
PlotSpectraBPIadj
}
