% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PlotSpectraInsensityStistics}
\alias{PlotSpectraInsensityStistics}
\title{Title}
\usage{
PlotSpectraInsensityStistics(
  mSet,
  imgName,
  format = "png",
  dpi = 72,
  width = NA
)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
