% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotTICs}
\alias{plotTICs}
\title{plotTICs}
\usage{
plotTICs(mSet = NULL, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{mSet}{mSet Object, should be processed by ImportMSData.}

\item{imgName}{Character, to name the imgName for the TIC plotting.}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png". (only works for web version)}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72. (only works for web version)}
}
\description{
plotTICs is used to plot the TIC of all files
}
