% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PlotXIC}
\alias{PlotXIC}
\title{Plot EIC}
\usage{
PlotXIC(featureNum, sample_labeled, Group_labeled, format, dpi, width, height)
}
\arguments{
\item{format}{Character, input the format of the image to create.}

\item{dpi}{Numeric, input the dpi of the image to create.}

\item{width}{Numeric, input the width of the image to create.}

\item{raw_data}{The object created using the ImportRawMSData function,
containing the raw MS data.}

\item{rt_mn}{Numeric, specify the minimum bound of the retention time range.}

\item{rt_mx}{Numeric, specify the maximum bound of the retention time range.}

\item{mz_mn}{Numeric, specify the minimum bound of the m/z range.}

\item{mz_mx}{Numeric, specify the maximum bound of the m/z range.}

\item{aggreg}{Character, if "sum", creates a total ion chromatogram.
If "max", creates a base peak chromatogram. By default it is set
to "sum".}
}
\description{
This functionn creates an extracted ion chromatogram (EIC) for a specific
m/z and retention time. This is used for quality-control of raw m/s data.
}
