% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{running.plan}
\alias{running.plan}
\title{running.plan}
\usage{
running.plan(plan = NULL, ...)
}
\arguments{
\item{plan}{ResummingPlan object. The object is generated by 'InitializaPlan' function.}

\item{...}{Multiple Processing commands can be input here.}
}
\description{
running.plan
}
\examples{
library(OptiLCMS);
plan <- InitializaPlan("raw_opt")
plan <- running.plan(plan,
                     data_folder_QC <- paste0(dataDir,"/QC"),
                     mSet <- PerformROIExtraction(datapath = data_folder_QC, rt.idx = 0.5, 
                                                  plot = F, rmConts = F, 
                                                  running.controller = rc),
                     param_initial <- SetPeakParam(),
                     best_parameters <- PerformParamsOptimization(mSet = mSet, param_initial, 
                                                                  ncore = 1, 
                                                                  running.controller = rc),
                     param <- best_parameters,
                     plotSettings1 <- SetPlotParam(Plot=T),
                     plotSettings2 <- SetPlotParam(Plot=T),
                     mSet <- ImportRawMSData(mSet = mSet, foldername = dataDir, 
                                             plotSettings = plotSettings1, 
                                             running.controller = rc),
                     mSet <- PerformPeakProfiling(mSet = mSet, Params = param, 
                                                  plotSettings = plotSettings2, 
                                                  running.controller = rc),
                     annParams <- SetAnnotationParam(polarity = 'negative', 
                                                     mz_abs_add = 0.025),
                     mSet <- PerformPeakAnnotation(mSet = mSet, 
                                                   annotaParam = annParams, ncore =1, 
                                                   running.controller = rc),
                     mSet <- FormatPeakList(mSet = mSet, annParams, 
                                            filtIso =F, filtAdducts = FALSE, 
                                            missPercent = 1));
}
\seealso{
\code{\link{ExecutePlan}} for the this resumable running pipeline.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
