% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{SetAnnotationParam}
\alias{SetAnnotationParam}
\title{Set annotation parameters}
\usage{
SetAnnotationParam(
  polarity = "positive",
  perc_fwhm = 0.6,
  mz_abs_iso = 0.005,
  max_charge = 2,
  max_iso = 2,
  corr_eic_th = 0.85,
  mz_abs_add = 0.001,
  adducts = NULL
)
}
\arguments{
\item{polarity}{Character, specify the polarity of the MS instrument. Either
"negative" or "positive".}

\item{perc_fwhm}{Numeric, set the percentage of the width of the FWHM for peak grouping.
Default is set to 0.6.}

\item{mz_abs_iso}{Numeric, set the allowed variance for the search (for isotope annotation).
The default is set to 0.005.}

\item{max_charge}{Numeric, set the maximum number of the isotope charge. For example,
the default is 2, therefore the max isotope charge is 2+/-.}

\item{max_iso}{Numeric, set the maximum number of isotope peaks. For example, the default
is 2, therefore the max number of isotopes per peaks is 2.}

\item{corr_eic_th}{Numeric, set the threshold for intensity correlations across samples.
Default is set to 0.85.}

\item{mz_abs_add}{Numeric, set the allowed variance for the search (for adduct annotation).
The default is set to 0.001.}

\item{adducts}{Character, specify the adducts based on your instrument settings.}
}
\value{
will return a annotation parameter set for following annotation steps
}
\description{
This function sets the parameters for peak annotation.
}
\examples{
##' Set peak annotation parameters
annParams <- SetAnnotationParam(polarity = 'positive',
                                mz_abs_add = 0.035);

##' Please check the example of PerformPeakProfiling 
##' and ExcutePlan for the whole running pipeline.
}
\seealso{
\code{\link{ExecutePlan}} and \code{\link{PerformPeakProfiling}} for the whole pipeline.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
