% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_setting.R
\name{SetGlobalParallel}
\alias{SetGlobalParallel}
\title{SetGlobalParallel}
\usage{
SetGlobalParallel(ncore = 1)
}
\arguments{
\item{ncore}{Numeric, used to set the global core numbers, default is 1}
}
\description{
SetGlobalParallel used to set the global core numbers
}
\examples{
SetGlobalParallel(1)
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
