% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{sphingo_chain_mass}
\alias{sphingo_chain_mass}
\title{Sphingolipid Base Chain Mass}
\usage{
sphingo_chain_mass(carbon, db_bond)
}
\arguments{
\item{carbon}{Integer. Carbon count}

\item{db_bond}{Integer. Double bond count}
}
\value{
Numeric. Sphingoid chain neutral mass
}
\description{
Calculates sphingoid base chain neutral mass.
C# equivalent: LipidMsmsCharacterizationUtility.SphingoChainMass
}
\keyword{internal}
