% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_import.R
\name{UpdateRawfiles}
\alias{UpdateRawfiles}
\title{UpdateRawfiles}
\usage{
UpdateRawfiles(mSet, filesIncluded = NULL)
}
\arguments{
\item{mSet}{mSet objects generated with \"mSet<-InitDataObjects(\"spec\", \"raw\", FALSE)\";}

\item{filesIncluded}{filesIncluded is a vector containing the files' paths for the following processing;}
}
\description{
Update the Raw spectra included for Processing. All wrong format and uncentroided files will be filtered.
NOTE: this function is only effective before data import stage.
}
\examples{
## load googledrive package to download example data
# library("googledrive");
# data_folder_Sample <- "Raw_data_example"
# temp <- tempfile(fileext = ".zip");
## Please authorize the package to download the data from web
# dl <- drive_download(as_id("1CjEPed1WZrwd5T3Ovuic1KVF-Uz13NjO"), path = temp, overwrite = TRUE);
# out <- unzip(temp, exdir = data_folder_Sample);
# out;
# library(OptiLCMS);
# mSet<-InitDataObjects("spec", "raw", FALSE);
## include only two samples CD_SM-77FXR.mzML and CD_SM-6KUCT.mzML for data import.
# mSet<-UpdateRawfiles(mSet, c("Raw_data_example/CD/CD_SM-77FXR.mzML", "Raw_data_example/CD/CD_SM-6KUCT.mzML"))
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
