% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{updateRawSpectraParam}
\alias{updateRawSpectraParam}
\title{updateRawSpectraParam}
\usage{
updateRawSpectraParam(Params)
}
\arguments{
\item{Params}{object generated by SetPeakParams function.}
}
\description{
updateRawSpectraParam
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
Mcgill University
License: GNU GPL (>= 2)
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
}
