% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{updateRawSpectraPath}
\alias{updateRawSpectraPath}
\title{updateRawSpectraPath}
\usage{
updateRawSpectraPath(mSet, newPath)
}
\arguments{
\item{mSet}{mSet object generated by ImportRawMSData or the following functions.}

\item{newPath}{Character vector, a character vector specify the absolute path of the new raw spectra files.}
}
\description{
updateRawSpectraPath
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
}
