% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{CentroidCheck}
\alias{CentroidCheck}
\title{Verify the data is centroid or not}
\usage{
CentroidCheck(filename)
}
\arguments{
\item{filename}{single file name, should contain the absolute path}
}
\description{
Verify the data is centroid or not
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
