% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_import.R
\name{CentroidCheck}
\alias{CentroidCheck}
\title{CentroidCheck}
\usage{
CentroidCheck(filename)
}
\arguments{
\item{filename}{single file name, should contain the absolute path}
}
\description{
Verify the data is centroid or not
}
\examples{
 
DataFiles <- dir(system.file("mzData", package = "mtbls2"), full.names = TRUE,
                 recursive = TRUE)[c(10:12, 14:16)]
sapply(DataFiles, CentroidCheck)
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
