% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{controller.modifier}
\alias{controller.modifier}
\title{operators_1 : PerformDataTrimming ~ PerformParamsOptimization;
operators_2 : PerformParamsOptimization ~ peak_profiling;  #'
operators_3 : PerformPeakProfiling ~ PerformPeakAnnotation;  #'
operators_4 : ;
others_1 c1 : Control Optimization or not;
others_1 c2 : ;
others_1 c3 : ;
others_1 c4 : .}
\usage{
controller.modifier(new_command_set, last_command_set, plan)
}
\description{
operators_1 : PerformDataTrimming ~ PerformParamsOptimization;
operators_2 : PerformParamsOptimization ~ peak_profiling;  #'
operators_3 : PerformPeakProfiling ~ PerformPeakAnnotation;  #'
operators_4 : ;
others_1 c1 : Control Optimization or not;
others_1 c2 : ;
others_1 c3 : ;
others_1 c4 : .
}
