% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{creatPeakTable}
\alias{creatPeakTable}
\title{creatPeakTable}
\usage{
creatPeakTable(mSet)
}
\arguments{
\item{mSet}{mSet object, usually generated by 'PerformPeakAnnotation' here.}
}
\description{
creatPeakTable
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
