% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{ExecutePlan}
\alias{ExecutePlan}
\title{ExecutePlan}
\usage{
ExecutePlan(plan = NULL)
}
\arguments{
\item{plan}{ResummingPlan object. The object is generated by running.plan() function.}
}
\description{
ExecutePlan
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
