% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{Export.Annotation}
\alias{Export.Annotation}
\title{Export.Annotation}
\usage{
Export.Annotation(mSet = NULL, path = getwd())
}
\arguments{
\item{mSet}{mSet object, processed by FormatPeakList.}

\item{path}{character, used to specify the path for result rds and csv file. Default is the working directory.}
}
\value{
will save annotated_peaklist.rds and annotated_peaklist.csv into working path
}
\description{
Export.Annotation is used to export the result of annotation
}
\examples{
data(mSet)
Export.Annotation(mSet, path = tempdir())
# delete the exported files from the tempdir with unlink
unlink(paste0(tempdir(),"/annotated_peaklist.csv"), recursive = TRUE, force = TRUE);
unlink(paste0(tempdir(),"/annotated_peaklist.rds"), recursive = TRUE, force = TRUE)
}
\seealso{
\code{\link{ExecutePlan}} and \code{\link{PerformPeakProfiling}} for the whole pipeline.
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca}, Jeff Xia \email{jeff.xia@mcgill.ca}
}
