% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{FormatMSnAnnotation}
\alias{FormatMSnAnnotation}
\title{FormatMSnAnnotation}
\usage{
FormatMSnAnnotation(mSet = NULL, topN = 5L, isLipidomics = FALSE)
}
\arguments{
\item{mSet}{}

\item{topN}{}

\item{isLipidomics}{}
}
\value{
mSet Object
}
\description{
FormatMSnAnnotation
}
