% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{FormatPeakList}
\alias{FormatPeakList}
\title{Format Peak List}
\usage{
FormatPeakList(
  mSet,
  annParams,
  filtIso = TRUE,
  filtAdducts = FALSE,
  missPercent = 0.75
)
}
\arguments{
\item{mSet}{The mSet object generated by the PerformPeakAnnotation function.}

\item{annParams}{The object created using the SetAnnotationParam function,
containing user's specified or default parameters for downstream
raw MS data pre-processing.}

\item{filtIso}{Logical, filter out all isotopes except for \verb{[M]}+ for
positive ion mode and \verb{[M]}- for negative ion mode. By default it is
set to true.}

\item{filtAdducts}{Logical, filter out all adducts except \verb{[M+H]}+ for
positive ion more and \verb{[M-H]}- for negative ion mode. By default it is set to false.}

\item{missPercent}{Numeric, specify the threshold to remove features
missing in X\\% of samples. For instance, 0.5 specifies to remove features
that are missing from 50\\% of all samples per group. Method is only valid
when there are two groups.}
}
\description{
This function formats the CAMERA output to a usable format for OptiLCMS.
}
\examples{
## load googledrive package to download example data
# library("googledrive");

## Set data folder
# data_folder_Sample <- "~/Data_IBD";
# data_folder_QC <- "~/Data_IBD/QC";
# temp <- tempfile(fileext = ".zip");

## Please authorize the package to download the data from web
# dl <- drive_download(as_id("1CjEPed1WZrwd5T3Ovuic1KVF-Uz13NjO"), path = temp, overwrite = TRUE);
# out <- unzip(temp, exdir = data_folder_Sample);
# out;
#### Running as regular procedure: step by step
## Extract ROI for parameters' optimization
# mSet <- PerformROIExtraction(datapath = data_folder_QC, rt.idx = 0.95, plot = F, rmConts = F);
## Perform the optimization
# best_parameters <- PerformParamsOptimization(mSet = mSet, SetPeakParam(), ncore = 4);
## Perform data import of all samples
# mSet <- ImportRawMSData(mSet = mSet, foldername = data_folder_Sample, 
#                         plotSettings = SetPlotParam(Plot=T));
## Perform peak profiling
# mSet <- PerformPeakProfiling(mSet = mSet, Params = param, plotSettings = SetPlotParam(Plot=T));
## Set annotation parameters
# annParams <- SetAnnotationParam(polarity = 'negative', mz_abs_add = 0.025);
## Perform peak annotation
# mSet <- PerformPeakAnnotation(mSet = mSet, annotaParam = annParams, ncore =1);
## Format the peak table
# maPeaks <- FormatPeakList(mSet = mSet, annParams, filtIso =F, 
#                           filtAdducts = FALSE,missPercent = 1);
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
