% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{GaussModel}
\alias{GaussModel}
\title{GaussModel}
\usage{
GaussModel(x, mu, sigma, h)
}
\arguments{
\item{x}{a numeric vector of values at which to evaluate the model}

\item{mu}{mean  of the distribution function}

\item{sigma}{standard deviation of the distribution fuction}

\item{h}{height of the distribution function}
}
\description{
GaussModel
}
\examples{
ints<- c(c(1:5,5:1))
##nls(y ~ GaussModel(x, mu, sigma, h), 
##                   data.frame(x = 1:length(ints), y = ints))
}
\references{
Smith, C.A., Want, E.J., O'Maille, G., Abagyan,R., Siuzdak, G. (2006).
"XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear
peak alignment, matching and identification." Analytical Chemistry, 78, 779-787.
}
