% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LipidMsmsCharacterization.R
\name{judge_if_pbtoh}
\alias{judge_if_pbtoh}
\title{Phosphatidylbutanol (PBtOH) Classification}
\usage{
judge_if_pbtoh(
  ms_scan_prop,
  ms2_tolerance,
  theoretical_mz,
  total_carbon,
  total_double_bond,
  min_sn_carbon,
  max_sn_carbon,
  min_sn_double_bond,
  max_sn_double_bond,
  adduct
)
}
\value{
PBtOH level 2 annotation (acyl_count=2) or NULL if criteria not met
}
\description{
Identifies PBtOH lipids using 2D iteration over acyl chains.
}
\details{
\strong{Positive Mode \link{M+NH4}+:}
Seeks diagnostic m/z = theoreticalMz - 17.026549 (\link{M+H-NH3}+, threshold 1).
2D iteration detecting acyl_cation - electron ions (intensity 0.1).
Requires foundCount = 2.

\strong{Negative Mode \link{M-H}-:}
Seeks phosphate head m/z 153.03221938 (\link{C4H10O4P}-, threshold 1).
2D iteration detecting \link{FA-H}- ions (intensity 5.0). Requires foundCount = 2.
}
\keyword{internal}
