% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PeakPicking_Massifquant_slave}
\alias{PeakPicking_Massifquant_slave}
\title{PeakPicking_Massifquant_slave}
\usage{
PeakPicking_Massifquant_slave(x, param)
}
\description{
PeakPicking_Massifquant_slave
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
