% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformDBSearchingBatch}
\alias{PerformDBSearchingBatch}
\title{PerformDBSearchingBatch}
\usage{
PerformDBSearchingBatch(
  mSet = NULL,
  ppm1 = 5,
  ppm2 = 15,
  rt_tol = 0,
  database_path = "",
  use_rt = FALSE,
  enableNL = FALSE,
  NLdatabase_path = NULL,
  ncores = 1,
  useEntropy = FALSE,
  databaseOptions = "all"
)
}
\arguments{
\item{mSet}{mSet Object contains raw spectral data after results consensus from \emph{PerformSpectrumConsenus}}

\item{ppm1}{numeric, ppm value of m/z for precursours;}

\item{ppm2}{numeric, ppm value of m/z for ms/ms fragments matching;}

\item{rt_tol}{numeric, retention time tolerance, in seconds. Only effective when use_rt is TRUE;}

\item{database_path}{character, specify the path of database (.sqlite format);}

\item{use_rt}{logical, to use retention time if TRUE;}

\item{enableNL}{logical, to enable use Neutral Loss matching for unmatched features if TRUE;}

\item{NLdatabase_path}{path of neutral loss database. Must be specified to a valid neutral loss database when enableNL is TRUE.}

\item{ncores}{}

\item{databaseOptions}{Options of database for searching.}
}
\value{
mSet Object
}
\description{
PerformDBSearchingBatch
}
