% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformDBSearchingBatch}
\alias{PerformDBSearchingBatch}
\title{PerformDBSearchingBatch}
\usage{
PerformDBSearchingBatch(
  mSet = NULL,
  ppm1 = 5,
  ppm2 = 15,
  rt_tol = 0,
  database_path = "",
  use_rt = FALSE,
  enableNL = FALSE,
  ncores = 1,
  useEntropy = FALSE
)
}
\arguments{
\item{mSet}{}

\item{ppm1}{}

\item{ppm2}{}

\item{rt_tol}{}

\item{database_path}{}

\item{use_rt}{}

\item{enableNL}{}

\item{ncores}{}
}
\value{
mSet Object
}
\description{
PerformDBSearchingBatch
}
\examples{
to add
}
