% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformDIADeconvolution}
\alias{PerformDIADeconvolution}
\title{PerformDIADeconvolution}
\usage{
PerformDIADeconvolution(
  mSet = NULL,
  min_width = 5,
  ppm2,
  sn = 12,
  span = 0.3,
  filtering = 2000,
  ncores = 1L
)
}
\arguments{
\item{mSet}{}

\item{min_width}{}

\item{ppm2}{}

\item{sn}{}

\item{span}{}

\item{filtering}{}

\item{ncores}{}
}
\value{
mSet Object
}
\description{
PerformDIADeconvolution
}
\examples{
to add
}
