% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Infusion_function.R
\name{PerformInusionFeatureDection}
\alias{PerformInusionFeatureDection}
\title{PerformInusionFeatureDection}
\usage{
PerformInusionFeatureDection(
  mSet,
  method = "cwt",
  noise = NA,
  snthresh = 3,
  ppm = 10,
  scales = c(1, 4, 9),
  ncores = 1L
)
}
\arguments{
\item{mSet}{}

\item{method}{}

\item{noise}{}

\item{snthresh}{}

\item{ppm}{}

\item{scales}{}

\item{ncores}{}

\item{minFraction}{}

\item{minSamples}{}
}
\value{
mSet Object
}
\description{
PerformInusionFeatureDection
}
\examples{
todo
}
