% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakAlignment}
\alias{PerformPeakAlignment}
\title{PerformPeakAlignment}
\usage{
PerformPeakAlignment(mSet)
}
\arguments{
\item{mSet}{the mSet object generated by PerformPeakPicking function.}
}
\value{
will return an mSet object with peak aligned done
}
\description{
PerformPeakAlignment
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
mSet <- PerformPeakAlignment(mSet);
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
