% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakPicking}
\alias{PerformPeakPicking}
\title{PerformPeakPicking}
\usage{
PerformPeakPicking(mSet, BPPARAM = bpparam())
}
\arguments{
\item{BPPARAM}{parallel method used for data processing. Default is bpparam().}

\item{object}{the raw data object read by ImportRawMSData function.}
}
\description{
PerformPeakPicking
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
