% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{PerformPeakPicking}
\alias{PerformPeakPicking}
\title{PerformPeakPicking}
\usage{
PerformPeakPicking(mSet, BPPARAM = bpparam())
}
\arguments{
\item{mSet}{the raw data object read by ImportRawMSData function.}

\item{BPPARAM}{parallel method used for data processing. Default is bpparam(). Optional.}
}
\description{
PerformPeakPicking
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
