% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformResultsExport}
\alias{PerformResultsExport}
\title{PerformResultsExport}
\usage{
PerformResultsExport(
  mSet = NULL,
  type = 0L,
  topN = 10L,
  ncores = 1L,
  lipids = F
)
}
\arguments{
\item{mSet}{}

\item{type}{}

\item{topN}{}

\item{ncores}{}

\item{lipids}{}
}
\value{
mSet Object
}
\description{
PerformResultsExport
}
