% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSn_processing.R
\name{PerformSpectrumConsenus}
\alias{PerformSpectrumConsenus}
\title{PerformSpectrumConsenus}
\usage{
PerformSpectrumConsenus(
  mSet = NULL,
  ppm2,
  concensus_fraction = 0.5,
  database_path = "",
  use_rt = FALSE,
  user_dbCorrection = TRUE,
  useEntropy = FALSE
)
}
\arguments{
\item{mSet}{}

\item{ppm2}{}

\item{concensus_fraction}{}

\item{database_path}{}

\item{use_rt}{}

\item{user_dbCorrection}{}
}
\value{
mSet Object
}
\description{
PerformSpectrumConsenus
}
\examples{
to add
}
