% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotMSfeature}
\alias{plotMSfeature}
\title{plotMSfeature}
\usage{
plotMSfeature(mSet, FeatureNM, dpi = 72, format = "png")
}
\arguments{
\item{mSet}{mSet Object, should be processed aby 'PerformPeakProfiling'.}

\item{FeatureNM}{Numeric, feature number in the feature table.}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72. (only works for web version)}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png". (only works for web version)}
}
\description{
plotMSfeature is used to plot the feature intensity of different groups
}
\examples{
library(OptiLCMS)
data(mSet);
plotMSfeature (mSet, 1); # Here is only one group
}
