% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotMSfeature}
\alias{plotMSfeature}
\title{plotMSfeature}
\usage{
plotMSfeature(mSet = NULL, FeatureNM = 1, dpi = 72, format = "png", width = NA)
}
\arguments{
\item{mSet}{mSet Object, should be processed aby 'PerformPeakProfiling'.}

\item{FeatureNM}{Numeric, feature number in the feature table.}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72. (only works for web version)}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png". (only works for web version)}

\item{width}{Numeric, width of the figure (default is NA, usually set it as 6~12)}
}
\value{
will return a figure of ms stats
}
\description{
plotMSfeature is used to plot the feature intensity of different groups
}
\examples{
data(mSet);
newPath <- dir(system.file("mzData", package = "mtbls2"),
               full.names = TRUE, recursive = TRUE)[c(10, 11, 12)]
mSet <- updateRawSpectraPath(mSet, newPath);
plotMSfeature (mSet, 1); # Here is only one group
}
