% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{plotSingleTIC}
\alias{plotSingleTIC}
\title{plotSingleTIC}
\usage{
plotSingleTIC(mSet = NULL, filename, imagename)
}
\arguments{
\item{mSet}{mSet Object, should be processed by ImportMSData.}

\item{filename}{Character, to give the filename for the TIC plotting.}

\item{imagename}{Character, to give the filename of the TIC plotted. (only works for web version)}
}
\description{
plotSingleTIC is used to plot the TIC of a certain spectra
}
\examples{
library(OptiLCMS)
data(mSet);
plotSingleTIC(mSet, "MSpos-Ex2-Col0-48h-Ag-2_1-A,3_01_9829.mzData")
}
