% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PlotSpectraPCA}
\alias{PlotSpectraPCA}
\title{Title}
\usage{
PlotSpectraPCA(mSet, imgName, format = "png", dpi = 72, width = NA)
}
\arguments{
\item{width}{}
}
\value{

}
\description{
Title
}
