% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_visulization.R
\name{PlotXIC}
\alias{PlotXIC}
\title{PlotXIC/EIC}
\usage{
PlotXIC(
  mSet = NULL,
  featureNum,
  sample_labeled,
  Group_labeled,
  format,
  dpi,
  width,
  height
)
}
\arguments{
\item{mSet}{mSet Object. Should contain the spectra processing result.}

\item{featureNum}{Numeric, Feature number in the feature table.}

\item{sample_labeled}{Logical, whether to lable the sample name.}

\item{Group_labeled}{Logical, whether to lable the group name.}

\item{format}{Character, to give the format of BPI figures ploted. Can be "jpeg", "png", "pdf", "svg",
"tiff" or "ps". Default is "png".}

\item{dpi}{Numeric, to define the dpi of the figures. Default is 72.}

\item{width}{Numeric, to define the width of the figure.}

\item{height}{Numeric, to define the height of the figure.}
}
\description{
This functionn creates an extracted ion chromatogram (EIC) for a specific
m/z and retention time. This is used for quality-control of raw m/s data.
}
\examples{
library(OptiLCMS)
data(mSet);
PlotXIC(mSet, 1, TRUE, TRUE);
}
