% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra_Utils.R
\name{RT.Adjust_peakGroup}
\alias{RT.Adjust_peakGroup}
\title{RT.Adjust_peakGroup}
\usage{
RT.Adjust_peakGroup(
  peaks,
  peakIndex,
  rtime,
  minFraction = 0.9,
  extraPeaks = 1,
  smooth = c("loess", "linear"),
  span = 0.2,
  family = c("gaussian", "symmetric"),
  peakGroupsMatrix = matrix(ncol = 0, nrow = 0),
  subset = integer(),
  subsetAdjust = c("average", "previous")
)
}
\description{
RT.Adjust_peakGroup
}
\references{
Smith, C.A. et al. 2006. {Analytical Chemistry}, 78, 779-787
}
\author{
Zhiqiang Pang, Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
