% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resume_functions.R
\name{running.plan}
\alias{running.plan}
\title{running.plan}
\usage{
running.plan(plan = NULL, ...)
}
\arguments{
\item{plan}{ResummingPlan object. The object is generated by 'InitializaPlan' function.}

\item{...}{Multiple Processing commands can be input here.}
}
\description{
running.plan
}
\author{
Zhiqiang Pang \email{zhiqiang.pang@mail.mcgill.ca} Jeff Xia \email{jeff.xia@mcgill.ca}
Mcgill University
License: GNU GPL (>= 2)
}
