% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform_functions.R
\name{SetAnnotationParam}
\alias{SetAnnotationParam}
\title{Set annotation parameters}
\usage{
SetAnnotationParam(
  polarity = "positive",
  perc_fwhm = 0.6,
  mz_abs_iso = 0.005,
  max_charge = 2,
  max_iso = 2,
  corr_eic_th = 0.85,
  mz_abs_add = 0.001
)
}
\arguments{
\item{polarity}{Character, specify the polarity of the MS instrument. Either
"negative" or "positive".}

\item{perc_fwhm}{Numeric, set the percentage of the width of the FWHM for peak grouping.
Default is set to 0.6.}

\item{mz_abs_iso}{Numeric, set the allowed variance for the search (for isotope annotation).
The default is set to 0.005.}

\item{max_charge}{Numeric, set the maximum number of the isotope charge. For example,
the default is 2, therefore the max isotope charge is 2+/-.}

\item{max_iso}{Numeric, set the maximum number of isotope peaks. For example, the default
is 2, therefore the max number of isotopes per peaks is 2.}

\item{corr_eic_th}{Numeric, set the threshold for intensity correlations across samples.
Default is set to 0.85.}

\item{mz_abs_add}{Numeric, set the allowed variance for the search (for adduct annotation).
The default is set to 0.001.}
}
\description{
This function sets the parameters for peak annotation.
}
\examples{
## load googledrive package to download example data
# library("googledrive");

## Set data folder
# data_folder_Sample <- "~/Data_IBD";
# data_folder_QC <- "~/Data_IBD/QC";
# temp <- tempfile(fileext = ".zip");

## Please authorize the package to download the data from web
# dl <- drive_download(as_id("1CjEPed1WZrwd5T3Ovuic1KVF-Uz13NjO"), path = temp, overwrite = TRUE);
# out <- unzip(temp, exdir = data_folder_Sample);
# out;
#### Running as regular procedure: step by step
## Extract ROI for parameters' optimization
# mSet <- PerformROIExtraction(datapath = data_folder_QC, rt.idx = 0.95, plot = F, rmConts = F);
## Perform the optimization
# best_parameters <- PerformParamsOptimization(mSet = mSet, SetPeakParam(), ncore = 4);
## Perform data import of all samples
# mSet <- ImportRawMSData(mSet = mSet, foldername = data_folder_Sample, plotSettings = SetPlotParam(Plot=T));
## Perform peak profiling
# mSet <- PerformPeakProfiling(mSet = mSet, Params = param, plotSettings = SetPlotParam(Plot=T));
## Set annotation parameters
# annParams <- SetAnnotationParam(polarity = 'negative', mz_abs_add = 0.025);
## Perform peak annotation
# mSet <- PerformPeakAnnotation(mSet = mSet, annotaParam = annParams, ncore =1);
## Format the peak table
# maPeaks <- FormatPeakList(mSet = mSet, annParams, filtIso =F, filtAdducts = FALSE,missPercent = 1);
}
\author{
Jasmine Chong \email{jasmine.chong@mail.mcgill.ca},
Mai Yamamoto \email{yamamoto.mai@mail.mcgill.ca}, and Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
